+++ gridstatus/tests/test_eia.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import pytest
 
 import gridstatus
 
@@ -107,3 +108,47 @@ def test_other_dataset():
 
     assert df.columns.tolist() == cols
     assert df.shape[0] > 0
+
+
+@pytest.mark.slow
+def test_eia_grid_monitor():
+    eia = gridstatus.EIA()
+    cols = [
+        "Area Id",
+        "Area Name",
+        "Area Type",
+        "Interval Start",
+        "Interval End",
+        "Demand",
+        "Demand Forecast",
+        "Net Generation",
+        "Total Interchange",
+        "NG: COL",
+        "NG: NG",
+        "NG: NUC",
+        "NG: OIL",
+        "NG: WAT",
+        "NG: SUN",
+        "NG: WND",
+        "NG: UNK",
+        "NG: OTH",
+        "Positive Gen from COL",
+        "Positive Gen from NG",
+        "Positive Gen from OIL",
+        "Positive Generation",
+        "Consumed Electricity",
+        "CO2 Factor: COL",
+        "CO2 Factor: NG",
+        "CO2 Factor: OIL",
+        "CO2 Emissions: COL",
+        "CO2 Emissions: NG",
+        "CO2 Emissions: OIL",
+        "CO2 Emissions: Other",
+        "CO2 Emissions Generated",
+        "CO2 Emissions Imported",
+        "CO2 Emissions Exported",
+        "CO2 Emissions Consumed",
+    ]
+    df = eia.get_grid_monitor()
+
+    assert df.columns.tolist() == cols